+++ OpenRA.Mods.Common/Traits/World/ElevatedBridgeLayer.cs
@@ -78,14 +78,14 @@ namespace OpenRA.Mods.Common.Traits
 			return cellCenters[cell];
 		}
 
-		int ICustomMovementLayer.EntryMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
+		short ICustomMovementLayer.EntryMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
 		{
-			return ends.Contains(cell) ? 0 : int.MaxValue;
+			return (short) (ends.Contains(cell) ? 0 : short.MaxValue);
 		}
 
-		int ICustomMovementLayer.ExitMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
+		short ICustomMovementLayer.ExitMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
 		{
-			return ends.Contains(cell) ? 0 : int.MaxValue;
+			return (short) (ends.Contains(cell) ? 0 : short.MaxValue);
 		}
 
 		byte ICustomMovementLayer.GetTerrainIndex(CPos cell)
