+++ OpenRA.Mods.Common/Traits/World/JumpjetActorLayer.cs
@@ -87,16 +87,16 @@ namespace OpenRA.Mods.Common.Traits
 			return ti == null || ti.RampType == 0;
 		}
 
-		int ICustomMovementLayer.EntryMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
+		short ICustomMovementLayer.EntryMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
 		{
 			var jli = (JumpjetLocomotorInfo)li;
-			return ValidTransitionCell(cell, jli) ? jli.JumpjetTransitionCost : int.MaxValue;
+			return (short) (ValidTransitionCell(cell, jli) ? jli.JumpjetTransitionCost : short.MaxValue);
 		}
 
-		int ICustomMovementLayer.ExitMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
+		short ICustomMovementLayer.ExitMovementCost(ActorInfo a, LocomotorInfo li, CPos cell)
 		{
 			var jli = (JumpjetLocomotorInfo)li;
-			return ValidTransitionCell(cell, jli) ? jli.JumpjetTransitionCost : int.MaxValue;
+			return (short) (ValidTransitionCell(cell, jli) ? jli.JumpjetTransitionCost : short.MaxValue);
 		}
 
 		byte ICustomMovementLayer.GetTerrainIndex(CPos cell)
